+++ depthaware/models/ops/depthconv/src/depthconv_cuda_kernel.h
@@ -6,7 +6,7 @@
 torch::Tensor depthconv_im2col(
     torch::Tensor data_im,
     torch::Tensor data_depth,
-    const int alpha,
+    const double alpha,
     const int channels, const int height, const int width,
     const int ksize_h, const int ksize_w,
     const int pad_h, const int pad_w,
